cache: includes cpu features in cache keys
there's a discussion about "offline" compilation,
and if that happens, the CPU feature incompatibility
becomes the problem.

This hardens the cache against it by including the cpu
id in the file cache keys so that the content will be
used exactly on the same CPUs.

Note that this impacts only amd64 target.

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>