Fix O_DIRECTORY allows opening dir with write
This commit fixes a `path_open` behavior that allows a directory to be
`path_open`ed with the `rights::fd_write` flag.  Since Wazero considers
it to be an error to open a directory with the write flag normally, it
is safe to treat this as a bug.  Although rights/capability system is
being scrapped, we still use the `fd_read` and `fd_write` flag to
determine if write permission needs to be set.

fixes #2243

Signed-off-by: Yage Hu <me@huyage.dev>