Fix descriptor.Table buffer growth
Fixes InsertAt size calculation to avoid growing slices more than
needed. On insert the mask should calculate the index in the slice not
the number of elements in the table. Avoids a factor of 8 increase in
table size. Small optimisation to use the capcity of the slice to reduce
the average length whilst keeping allocations small. Clarified the grow
method.

Signed-off-by: Edward McFarlane <emcfarlane@buf.build>